requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
kafka-python>=2.0.2
boto3>=1.34.0
pyyaml>=6.0.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pillow>=10.2.0
pytesseract>=0.3.10
pdf2image>=1.17.0
mlflow>=2.10.0
scikit-learn>=1.4.0
pandas>=2.2.0
pyspark>=3.5.0
pydantic>=2.6.0
